refactor: move components to root directory
