+++ src/components/BoardNew/detail/ContentViewer.tsx
